Handle winner indeterminate case when only 1 game is uploaded
Previously this would trigger an error and thus a log excerpt when multiple
games were uploaded (i.e., a set). This issue was because the "Wireframe"
winner's idx was not bound in the accumulated player stats. This should also now
allow sets with non-determinate winner games to appear in the final GUI viz.