Beta fixes + small enhancements (#41)
* Handle winner indeterminate case when only 1 game is uploaded

Previously this would trigger an error and thus a log excerpt when multiple
games were uploaded (i.e., a set). This issue was because the "Wireframe"
winner's idx was not bound in the accumulated player stats. This should also now
allow sets with non-determinate winner games to appear in the final GUI viz.

* Handle all game end types (i.e., timeouts, quits)

* Handle the case where no valid games are found

In some cases no valid games are found on upload. In this case, the app should
error out with a proper error message.

This commit also cleans up some existing error handling in the app, which is
still a WIP.

* Handle the case for a single game where someone doesn't kill

When someone doesn't kill, they won't have a favorite kill move.
+ Relax cases for valid games

* Attempt to add game lengths

* Settle format for match length viz.

* WIP: Add new stat for avg pct kills

* Add avg kill% stat + final style for game length viz

* Use conversions for calculating moves